# Generalising bitwise simplifications