This PR generalises several lemmas on bitwise operators, streamlining the development.

The corresponding tests are already present.