@ehildenb 2024-01-10
Can you post performance numbers for the entire KEVM test-suite and the Kontrol test-suite please (comparing this branch to master). I suspect this may have a slowdown in some cases.