@PetarMax 2024-01-10
Running the entire performance tests locally for me is complicated. I will wait to see the CI numbers first.