@ehildenb 2024-01-10
There are instructions here for KEVM: https://github.com/runtimeverification/hs-backend-booster/issues/355

Similar instructions work for Kontrol.