@PetarMax 2024-01-11
These is the performance comparison for KEVM on all tests, it does not appear to be statistically significant overall, with the PR times being slightly faster than the master times:
```
 Test                                                                                           | Master-stats time | PR-stats time | (Master-stats/PR-stats) time 
------------------------------------------------------------------------------------------------+-------------------+---------------+------------------------------
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flipper-tau-pass-spec.k]               | 91.76             | 103.7         | 0.8848601735776278           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/bytes00-spec.k]                 | 58.69             | 62.14         | 0.9444802059864821           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-mului-pass-spec.k]                 | 67.28             | 71.21         | 0.9448111220334223           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-deny-diff-fail-rough-spec.k]       | 80.21             | 84.76         | 0.9463190184049078           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/storagevar01-spec.k]            | 51.75             | 54.43         | 0.950762447179864            
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/keccak00-spec.k]                | 44.57             | 46.86         | 0.9511310285958173           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-1-c-spec.k]      | 53.87             | 56.5          | 0.9534513274336283           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/staticloop00-a0lt10-spec.k]     | 49.16             | 51.45         | 0.9554907677356655           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transferFrom-success-2-spec.k]   | 65.0              | 67.78         | 0.9589849513130717           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-1-c-spec.k]  | 98.5              | 102.64        | 0.9596648480124708           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-failure-3-spec.k]  | 73.86             | 76.91         | 0.9603432583539202           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/encode-keccak00-spec.k]         | 41.16             | 42.48         | 0.9689265536723164           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dsvalue-peek-pass-rough-spec.k]        | 46.13             | 47.56         | 0.969932716568545            
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-addui-pass-spec.k]                 | 94.87             | 97.63         | 0.9717300010242754           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-1-a-spec.k]      | 77.48             | 79.38         | 0.9760644998740238           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/approve-spec.k]                  | 37.98             | 38.8          | 0.9788659793814433           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-failure-4-spec.k]  | 98.56             | 100.37        | 0.9819667231244396           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flipper-addu48u48-fail-rough-spec.k]   | 42.24             | 42.93         | 0.9839273235499651           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/dynamicarray00-spec.k]          | 35.4              | 35.95         | 0.9847009735744088           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/functional-spec.k]                    | 74.24             | 75.3          | 0.9859229747675963           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/functional-spec.k]                     | 573.16            | 581.06        | 0.9864041579182873           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-1-a-spec.k]  | 92.09             | 93.34         | 0.9866080994214699           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-2-b-spec.k]  | 87.59             | 88.65         | 0.9880428652002256           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/address00-spec.k]               | 35.2              | 35.61         | 0.9884863802302725           
 src/tests/integration/test_prove.py::test_pyk_prove[examples/solidity-code-spec.md]            | 152.47            | 154.24        | 0.988524377593361            
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/overflow00-overflow-spec.k]     | 33.04             | 33.37         | 0.9901108780341624           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-success-1-spec.k]  | 53.91             | 54.43         | 0.9904464449751974           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-subui-pass-spec.k]                 | 92.22             | 93.08         | 0.9907606360120327           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/approve-failure-spec.k]           | 47.32             | 47.76         | 0.9907872696817421           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dstoken-approve-fail-rough-spec.k]     | 82.5              | 83.21         | 0.9914673717101311           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transfer-success-1-spec.k]       | 46.95             | 47.35         | 0.9915522703273496           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/overflow00-nooverflow-spec.k]   | 35.6              | 35.89         | 0.9919197548063527           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/approve-success-spec.k]           | 50.46             | 50.87         | 0.9919402398270101           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-2-b-spec.k]      | 43.38             | 43.73         | 0.9919963411845416           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/requires01-a0gt0-spec.k]        | 36.98             | 37.24         | 0.993018259935553            
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dsvalue-read-pass-spec.k]              | 50.43             | 50.78         | 0.9931075226467113           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-2-c-spec.k]  | 46.03             | 46.29         | 0.9943832361201124           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/allowance-spec.k]                 | 41.78             | 42.01         | 0.9945251130683171           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/ecrecover00-sigvalid-spec.k]    | 93.01             | 93.5          | 0.9947593582887702           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/storagevar02-nooverflow-spec.k] | 46.6              | 46.84         | 0.9948761742100768           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-1-b-spec.k]      | 104.37            | 104.79        | 0.9959919839679359           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transferFrom-failure-1-spec.k]   | 74.54             | 74.82         | 0.996257685110933            
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flipper-ttl-pass-spec.k]               | 49.64             | 49.81         | 0.9965870307167235           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-muluu-pass-spec.k]                 | 89.74             | 90.0          | 0.9971111111111111           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transferFrom-success-1-spec.k]   | 63.86             | 64.04         | 0.9971892567145533           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/structarg00-spec.k]             | 37.87             | 37.95         | 0.9978919631093542           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/storagevar00-spec.k]            | 44.58             | 44.67         | 0.9979852249832102           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/staticarray00-spec.k]           | 36.48             | 36.49         | 0.9997259523157028           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/end-subuu-pass-spec.k]                 | 45.47             | 45.43         | 1.0008804754567466           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transferFrom-failure-2-spec.k]   | 70.23             | 70.12         | 1.0015687393040502           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-success-2-spec.k]        | 61.97             | 61.82         | 1.0024263992235523           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/balanceOf-spec.k]                 | 40.29             | 40.18         | 1.0027376804380288           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-success-1-spec.k]        | 76.98             | 76.71         | 1.0035197497066877           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-success-2-spec.k]  | 71.98             | 71.58         | 1.0055881531153954           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/cat-file-addr-pass-rough-spec.k]       | 48.83             | 48.46         | 1.0076351630210483           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-2-a-spec.k]      | 104.89            | 104.07        | 1.0078793120015375           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/totalSupply-spec.k]              | 30.28             | 30.04         | 1.007989347536618            
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/balanceOf-spec.k]                | 34.88             | 34.6          | 1.0080924855491329           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-subui-pass-rough-spec.k]           | 105.45            | 104.48        | 1.0092840735068913           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/storagevar03-spec.k]            | 48.52             | 48.05         | 1.0097814776274716           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dsvalue-read-pass-summarize-spec.k]    | 46.4              | 45.9          | 1.0108932461873639           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-1-d-spec.k]  | 57.41             | 56.75         | 1.0116299559471364           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transfer-success-2-spec.k]       | 45.4              | 44.85         | 1.0122630992196209           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transfer-failure-2-spec.k]       | 48.6              | 47.93         | 1.0139787189651575           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-failure-1-spec.k]  | 81.6              | 80.39         | 1.0150516233362357           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/allowance-spec.k]                | 38.25             | 37.68         | 1.0151273885350318           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-slip-pass-rough-spec.k]            | 140.04            | 137.76        | 1.0165505226480835           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-move-diff-rough-spec.k]            | 125.3             | 123.24        | 1.0167153521583903           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/totalSupply-spec.k]               | 37.09             | 36.4          | 1.018956043956044            
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-success-2-spec.k]    | 76.34             | 74.77         | 1.02099772636084             
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/cat-exhaustiveness-spec.k]             | 228.98            | 224.26        | 1.0210469990189959           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flipper-bids-pass-rough-spec.k]        | 122.45            | 119.69        | 1.023059570557273            
 src/tests/integration/test_prove.py::test_pyk_prove[examples/erc20-spec.md]                    | 206.74            | 201.9         | 1.0239722634967805           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-subui-fail-rough-spec.k]           | 64.79             | 63.27         | 1.0240240240240241           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dai-symbol-pass-spec.k]                | 42.76             | 41.66         | 1.0264042246759482           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flopper-file-pass-rough-spec.k]        | 367.81            | 358.18        | 1.0268859232787984           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/dai-adduu-fail-rough-spec.k]           | 38.71             | 37.54         | 1.031166755460842            
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/requires01-a0le0-spec.k]        | 150.25            | 145.55        | 1.0322913088285812           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-2-a-spec.k]  | 67.29             | 65.11         | 1.0334818000307173           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/encodepacked-keccak01-spec.k]   | 40.43             | 39.11         | 1.0337509588340579           
 src/tests/integration/test_prove.py::test_pyk_prove[functional/evm-int-simplifications-spec.k] | 157.68            | 152.12        | 1.0365500920326058           
 src/tests/integration/test_prove.py::test_pyk_prove[bihu/forwardToHotWallet-failure-2-spec.k]  | 62.36             | 60.02         | 1.0389870043318894           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-failure-1-b-spec.k]  | 149.86            | 143.89        | 1.041490027104038            
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/flopper-cage-pass-spec.k]              | 52.84             | 50.67         | 1.0428261298598778           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transferFrom-success-1-spec.k]    | 92.22             | 88.11         | 1.0466462376574737           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/storagevar02-overflow-spec.k]   | 47.54             | 45.42         | 1.0466754733597534           
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/transfer-failure-1-spec.k]       | 169.53            | 161.82        | 1.0476455320726734           
 src/tests/integration/test_prove.py::test_pyk_prove[mcd/vat-addui-fail-rough-spec.k]           | 75.78             | 71.89         | 1.0541104465155098           
 src/tests/integration/test_prove.py::test_pyk_prove[examples/storage-spec.md]                  | 159.07            | 149.39        | 1.0647968404846375           
 src/tests/integration/test_prove.py::test_pyk_prove[benchmarks/ecrecover00-siginvalid-spec.k]  | 90.5              | 84.61         | 1.0696135208604185           
 TOTAL                                                                                          | 7348.299999999998 | 7326.02       | 1.0030412147386982 
```
This is likely because the booster can easily discharge the concrete `requires` conditions when a rule doesn't apply. For example, in this rule
```
    rule X &Int (Y <<Int Z) => Y <<Int Z
      requires 0 <=Int X andBool X <Int pow256
       andBool 2 ^Int ( log2Int ( pow256 -Int X ) ) ==Int pow256 -Int X
       andBool Z ==Int log2Int ( pow256 -Int X )
       andBool 0 <=Int Y andBool Y <Int 2 ^Int ( 256 -Int Z )
      [simplification, concrete(X, Z), comm]
```
the first, second, and third condition are all fully concrete always.